refactor(translate): code cleanup - match file and command name - refine description - handle invalid/missing arguments - reformat display
